chore: start auth/session context
