--- src/web/src/app/tarefas/layout.tsx
@@ -1,18 +1,0 @@
-import "./globals.css";
-
-export const metadata = {
-  title: "tarefa",
-  description: "tarefas",
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang='en'>
-      <body>{children}</body>
-    </html>
-  );
-}
