+++ src/web/src/shared/auth/components/Auth.tsx
@@ -1,0 +1,20 @@
+'use client';
+
+import { useAuth } from "../context/AuthContext"
+import { PublicPage } from "./Public/PublicPage";
+
+export function Auth({ children }: { children: React.ReactNode }) {
+  const { session } = useAuth()
+
+  if (session.status === 'pending') {
+    return <h1>Loading...</h1>
+  }
+
+  if (session.status === 'unauthenticated') {
+    return <PublicPage />    
+  }
+
+  if (session.status === 'authenticated') {
+    return <>{children}</>
+  }
+}
