+++ src/web/src/shared/auth/components/Public/PublicPage.style.ts
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+export const AuthForm = styled.form`
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+
+  border: solid 1px #2F2F2F;
+  padding: 2rem;
+  border-radius: .5rem;
+  background: #010101;
+
+  div {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 0.5rem;    
+  }  
+`;
