+++ src/web/src/shared/components/Tarefas/components/TodoList.tsx
@@ -1,0 +1,31 @@
+import { ITask } from "@/app/types/tasks";
+import React from "react";
+import Task from "./Task";
+
+interface TodoListProps {
+  tasks: ITask[];
+}
+
+const TodoList: React.FC<TodoListProps> = ({ tasks }) => {
+  return (
+    <div className='overflow-x-auto'>
+      <table className='table w-full'>
+        {/* head */}
+        <thead>
+          <tr>
+            <th>Tarefas</th>
+            <th>Ações</th>
+            
+          </tr>
+        </thead>
+        <tbody>
+          {tasks.map((task) => (
+            <Task key={task.id} task={task} />
+          ))}
+        </tbody>
+      </table>
+    </div>
+  );
+};
+
+export default TodoList;
