+++ src/web/src/shared/components/Tarefas/layout.tsx
@@ -1,0 +1,18 @@
+import "./globals.css";
+
+export const metadata = {
+  title: "Tarefas",
+  description: "tarefas",
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang='en'>
+      <body>{children}</body>
+    </html>
+  );
+}
